requirements.txt
beautifulsoup4==4.7.1
cffi==1.11.5
cssselect2==0.2.1
defusedxml==0.5.0
lxml==4.3.0
Pillow==5.4.1
pycparser==2.19
soupsieve==1.7.2
tinycss2==0.6.1
urllib3==1.24.1
webencodings==0.5.1
xlrd==1.2.0
xlwt==1.3.0